Changed all doubles in trees to floats to reduce final size

+++ VandySkimmer/VandyClasses/EventInfo.h
@@ -7,17 +7,17 @@ class EventInfo {
         EventInfo() = default;
         virtual ~EventInfo() = default;
 
-        double get_z_vtx() {return z_vtx;};
-        void set_z_vtx(double vtx){ z_vtx = vtx; };
+        float get_z_vtx() {return z_vtx;};
+        void set_z_vtx(float vtx){ z_vtx = vtx; };
 
-        double get_z_vtx_truth() {return z_vtx_truth;};
-        void set_z_vtx_truth(double vtx){ z_vtx_truth = vtx; };        
+        float get_z_vtx_truth() {return z_vtx_truth;};
+        void set_z_vtx_truth(float vtx){ z_vtx_truth = vtx; };        
 
-        double get_ZDC_rate() {return ZDC_rate;};
-        void set_ZDC_rate(double rate){ ZDC_rate = rate; };
+        float get_ZDC_rate() {return ZDC_rate;};
+        void set_ZDC_rate(float rate){ ZDC_rate = rate; };
 
-        double get_cross_section() {return cross_section;};
-        void set_cross_section(double cs){ cross_section = cs; };
+        float get_cross_section() {return cross_section;};
+        void set_cross_section(float cs){ cross_section = cs; };
 
         bool is_dijet_event(int jetR_index) { return dijet_event[jetR_index]; };
         void set_dijet_event(int jetR_index, bool dijet){ dijet_event[jetR_index] = dijet; };
@@ -37,11 +37,11 @@ class EventInfo {
         float get_subleadTruth_pT(int jetR_index) {return subleadTruth_pT[jetR_index];};
         void set_subleadTruth_pT(int jetR_index, float pT){ subleadTruth_pT[jetR_index] = pT; };
     
-        double get_dijetDeltat(int jetR_index) { return dijetDeltat[jetR_index]; };
-        void set_dijetDeltat(int jetR_index, double dijet){ dijetDeltat[jetR_index] = dijet; };
+        float get_dijetDeltat(int jetR_index) { return dijetDeltat[jetR_index]; };
+        void set_dijetDeltat(int jetR_index, float dijet){ dijetDeltat[jetR_index] = dijet; };
        
-	double get_dijetDeltatTruth(int jetR_index) { return dijetDeltatTruth[jetR_index]; };
-        void set_dijetDeltatTruth(int jetR_index, double dijet){ dijetDeltatTruth[jetR_index] = dijet; };
+	float get_dijetDeltatTruth(int jetR_index) { return dijetDeltatTruth[jetR_index]; };
+        void set_dijetDeltatTruth(int jetR_index, float dijet){ dijetDeltatTruth[jetR_index] = dijet; };
 
         bool is_dijetDeltatPass(int jetR_index) { return dijetDeltatPass[jetR_index]; };
         void set_dijetDeltatPass(int jetR_index, bool dijet){ dijetDeltatPass[jetR_index] = dijet; };
@@ -57,22 +57,22 @@ class EventInfo {
 
     
     private:
-        double 	z_vtx{0.0};
-        double 	z_vtx_truth{0.0};
-        double 	ZDC_rate{0.0};
+        float 	z_vtx{0.0};
+        float 	z_vtx_truth{0.0};
+        float 	ZDC_rate{0.0};
          
 	bool 	dijet_event[4] = {false, false, false, false};
         bool 	dijetTruth_event[4] = {false, false, false, false};
 	
-	double	dijetDeltat[4] {-999, -999, -999, -999};
-	double 	dijetDeltatTruth[4] {-999, -999, -999, -999};
+	float	dijetDeltat[4] {-999, -999, -999, -999};
+	float 	dijetDeltatTruth[4] {-999, -999, -999, -999};
 	bool	dijetDeltatPass[4] = {false, false, false, false};
 	bool 	dijetDeltatTruthPass[4] = {false, false, false, false};
 
 	float 	dijetDeltaPhi[4] {-999, -999, -999, -999};
 	float 	dijetDeltaPhiTruth[4] {-999, -999, -999, -999};
 
-        double 	cross_section{0.0};
+        float 	cross_section{0.0};
        
        	float 	lead_pT[4] = {-999, -999, -999, -999};
         float 	sublead_pT[4] = {-999, -999, -999, -999};
